server/interview/requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
aiosignal==1.2.0
altgraph==0.17.2
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.5.2
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autobahn==22.5.1
Automat==20.2.0
beautifulsoup4==4.9.3
black==21.4b2
bleach==4.1.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
channels==3.0.5
channels-redis==3.4.0
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.13
django-allauth==0.47.0
django-cors-headers==3.13.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==22.1
django-markdownx==3.0.1
django-ninja==0.13.0
django-ratelimit==3.0.1
django-seed==0.2.2
django-sslserver==0.22
django-summernote==0.8.20.0
django-user-agents==0.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
Faker==8.4.0
frozenlist==1.2.0
future==0.18.2
geoip2==4.2.0
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
httpie==3.2.1
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.8.2
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
maxminddb==2.0.3
MouseInfo==0.1.3
msgpack==1.0.4
multidict==5.1.0
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector-python==8.0.28
mysqlclient==2.0.3
numpy==1.22.3
oauthlib==3.1.1
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pathspec==0.8.1
pefile==2021.9.3
Pillow==8.4.0
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.20
pydantic==1.8.2
PyGetWindow==0.0.9
Pygments==2.12.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.2
PyJWT==2.3.0
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PySide6==6.2.0
PySocks==1.7.1
python-dateutil==2.8.1
python-telegram-bot==13.7
python3-openid==3.2.0
PyTweening==1.0.3
pytz==2021.1
pywin32-ctypes==0.2.0
ratelimit==2.2.1
regex==2021.4.4
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rich==12.4.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
selenium==4.0.0
selenium-stealth==1.0.6
service-identity==21.1.0
shiboken6==6.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==3.10.0.0
tzlocal==2.1
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.6
user-agents==2.2.0
webencodings==0.5.1
wsproto==1.0.0
yarl==1.6.3
zipp==3.6.0
zope.interface==5.4.0
